refactor: group tests, add custom error class, clarify logic
