+++ lib/classes/AbstractOpenApiSpec.js
@@ -49,7 +49,7 @@ class OpenApiSpec {
       expectedResponse = this.findResponseDefinition(expectedResponse.$ref);
     }
     if (!expectedResponse) {
-      throw new Error(`No '${status}' response defined for endpoint '${actualRequest.method} ${this.findOpenApiPathMatchingRequest(actualRequest)}' in OpenAPI spec`);
+      throw new ValidationError('WRONG_STATUS', `No '${status}' response defined for endpoint '${actualRequest.method} ${this.findOpenApiPathMatchingRequest(actualRequest)}' in OpenAPI spec`);
     }
 
     return { [status]: expectedResponse };
@@ -72,7 +72,7 @@ class OpenApiSpec {
     const pathItemObject = this.getPathItem(openApiPath);
     const operationObject = pathItemObject[actualRequest.method.toLowerCase()];
     if (!operationObject) {
-      throw new Error(`No '${actualRequest.method}' method defined for path '${openApiPath}' in OpenAPI spec`);
+      throw new ValidationError('WRONG_METHOD', `No '${actualRequest.method}' method defined for path '${openApiPath}' in OpenAPI spec`);
     }
     return operationObject;
   }
@@ -89,9 +89,6 @@ class OpenApiSpec {
       expectedResStatus,
       actualResponse.getBodyForValidation(),
     );
-    if (validationError) {
-      validationError.actualResponse = actualResponse.summary();
-    }
     return validationError;
   }
