+++ lib/classes/OpenApi2Spec.js
@@ -1,5 +1,6 @@
 const utils = require('../utils');
 const AbstractOpenApiSpec = require('./AbstractOpenApiSpec');
+const ValidationError = require('./errors/ValidationError');
 
 const doesOpenApiPathMatchPathname = (openApiPath, pathname) => {
   const pathInColonForm = utils.convertOpenApiPathToColonForm(openApiPath);
@@ -12,7 +13,7 @@ class OpenApi2Spec extends AbstractOpenApiSpec {
       doesOpenApiPathMatchPathname(openApiPath, pathname)
     );
     if (!openApiPath) {
-      throw new Error(`No '${pathname}' path defined in OpenAPI spec`);
+      throw new ValidationError('MATCHES_NO_PATHS', `No '${pathname}' path defined in OpenAPI spec`);
     }
     return openApiPath;
   }
