+++ lib/classes/OpenApi3Spec.js
@@ -3,6 +3,7 @@ const url = require('url');
 
 const utils = require('../utils');
 const AbstractOpenApiSpec = require('./AbstractOpenApiSpec');
+const ValidationError = require('./errors/ValidationError');
 
 const serversPropertyNotProvidedOrIsEmptyArray = (spec) =>
   (!spec.hasOwnProperty('servers') || !spec.servers.length);
@@ -70,14 +71,14 @@ class OpenApi3Spec extends AbstractOpenApiSpec {
   findOpenApiPathMatchingPathname(pathname) {
     const matchingServerBasePaths = this.getMatchingServerBasePaths(pathname);
     if (!matchingServerBasePaths.length) {
-      throw new Error(`No server matching '${pathname}' path defined in OpenAPI spec`);
+      throw new ValidationError('MATCHES_NO_SERVERS', `No server matching '${pathname}' path defined in OpenAPI spec`);
     }
     const openApiPath = this.paths().find(openApiPath =>
       doesOpenApiPathMatchPathname(openApiPath, pathname, matchingServerBasePaths)
     );
     if (!openApiPath) {
       const matchingServerUrls = this.getMatchingServerUrls(pathname);
-      throw new Error(`No '${pathname}' path defined in OpenAPI spec. (Matches servers ${inspect(matchingServerUrls)} but no 'server/endpointPath' combinations)`);
+      throw new ValidationError('MATCHES_NO_PATHS', `No '${pathname}' path defined in OpenAPI spec. (Matches servers ${inspect(matchingServerUrls)} but no 'server/endpointPath' combinations)`);
     }
     return openApiPath;
   }
