+++ lib/classes/errors/ValidationError.js
@@ -1,0 +1,8 @@
+class ValidationError extends Error {
+  constructor(code, message) {
+    super(message);
+    this.code = code;
+  }
+}
+
+module.exports = ValidationError;
