+++ lib/index.js
@@ -29,6 +29,9 @@ module.exports = function (filepathOrObject) {
       const actualRequest = actualResponse.req;
 
       const validationError = openApiSpec.validateResponse(actualResponse);
+      if (validationError) {
+        validationError.actualResponse = actualResponse.summary();
+      }
 
       this.assert(
         !validationError,
