+++ test/unit/assertions/satisfySchemaInApiSpec/satisfySchemaInApiSpec.test.js
@@ -2,7 +2,7 @@ const chai = require('chai');
 const path = require('path');
 const { inspect } = require('util');
 
-const chaiResponseValidator = require('../../..');
+const chaiResponseValidator = require('../../../..');
 
 const openApiSpecsDir = path.resolve('test', 'resources', 'exampleOpenApiFiles', 'valid', 'satisfySchemaInApiSpec');
 const openApiSpecs = [
